week_2/book_store/api/serializers.py
from rest_framework.serializers import ModelSerializer

from .models import Author, Book


class BookSerializer(ModelSerializer):
    class Meta:
        model = Book
        fields = '__all__'


class AuthorSerializer(ModelSerializer):
    book = BookSerializer(many=True, read_only=True)

    class Meta:
        model = Author
        fields = '__all__'
