pyproject.toml
[tool.poetry]
name = "itk-tasks"
version = "0.1.0"
description = ""
authors = ["Sergey <one6sergey@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
aiohttp = "^3.11.11"
ruff = "^0.9.2"
gunicorn = "^23.0.0"
psycopg2-binary = "^2.9.10"
redis = "^5.2.1"
django = "^5.1.5"
djangorestframework = "^3.15.2"
django-filter = "^24.3"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"


[tool.ruff.lint]
select = [
    "E",  # pycodestyle errors
    "W",  # pycodestyle warnings
    "F",  # pyflakes
    "I",  # isort
    "B",  # flake8-bugbear
    "C4",  # flake8-comprehensions
    "UP",  # pyupgrade
]
ignore = [
    "E501",  # line too long, handled by black
    "B008",  # do not perform function calls in argument defaults
    "C901",  # too complex
    "W191",  # indentation contains tabs
    "W291",
]